fix m_element_t to float and little edit on lr update

+++ include/matrix.h
@@ -6,7 +6,7 @@
 
 //#define M_ASSERTS 
 
-typedef double m_element_t;
+typedef float m_element_t;
 typedef m_element_t  (*m_fn)(m_element_t);
 typedef struct matrix_t{
         size_t  rows;
